+++ app/(public)/(auth)/layout.tsx
@@ -1,7 +1,4 @@
-export default function AuthLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
+export default function AuthLayout(props: LayoutProps<'/'>) {
+  const { children } = props;
   return <>{children}</>;
 }